@@ -1,37 +1,46 @@
 {
   "name": "roman-numerals-converter",
   "version": "1.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
     "serve": "vite preview",
-    "lint:fix": "eslint ./src --ext .jsx,.js,.ts,.tsx --quiet --fix --ignore-path ./.gitignore",
-    "lint:format": "prettier  --loglevel warn --write \"./**/*.{js,jsx,ts,tsx,css,md,json}\" ",
+    "lint:fix": "eslint './src/**/*.{ts,tsx,js,jsx}' --quiet --fix --ignore-path ./.gitignore",
+    "lint:format": "prettier  --log-level warn --write \"./**/*.{js,jsx,ts,tsx,css,md,json}\" ",
     "lint": "yarn lint:format && yarn lint:fix ",
-    "type-check": "tsc"
+    "type-check": "tsc",
+    "test": "vitest"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "eslint": "^8.57.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.10.2",
-    "@typescript-eslint/parser": "^5.10.2",
+    "@testing-library/jest-dom": "^6.4.5",
+    "@testing-library/react": "^16.0.0",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
     "@vitejs/plugin-react": "^1.3.2",
-    "eslint": "^8.8.0",
+    "@vitest/ui": "^1.6.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-react": "^7.34.2",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "jest": "^29.7.0",
+    "jsdom": "^24.1.0",
     "pre-commit": "^1.2.2",
-    "prettier": "^2.5.1",
+    "prettier": "^3.3.1",
+    "tsx": "^4.15.1",
     "typescript": "^4.5.5",
-    "vite": "^2.9.12"
+    "vite": "^5.2.13",
+    "vitest": "^1.6.0",
+    "vitest-dom": "^0.1.1"
   },
   "pre-commit": "lint",
   "license": "MIT"
